requirements.txt
Django>=4.2
djangorestframework==3.14.0
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.10